Do not subclass socket from daemon directly, it messes up error messages
